Add :data to the timeline events api
